refactor: move shared utilities (rate-limiter, retry, cache) to @opuspopuli/common
Move RateLimiter, withRetry/RetryPredicates, and MemoryCache from
extraction-provider into common so all packages (including region
plugins) can reuse them. Extraction-provider re-exports for backwards
compatibility.
