+++ packages/common/src/providers/rate-limiting/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Rate Limiting Module
+ *
+ * Provides token bucket rate limiting for controlling request rates.
+ */
+
+export * from "./types.js";
+export * from "./rate-limiter.js";
