+++ packages/common/src/providers/retry/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Retry Module
+ *
+ * Provides retry logic with exponential backoff and jitter
+ * for handling transient failures in external service calls.
+ */
+
+export * from "./types.js";
+export * from "./retry.js";
