+++ packages/extraction-provider/src/cache/index.ts
@@ -1,7 +1,13 @@
 /**
  * Cache module exports
+ *
+ * Core cache types (ICache, CacheOptions, MemoryCache) are now provided
+ * by @opuspopuli/common. Re-exported here for backwards compatibility.
  */
-export * from "./cache.interface.js";
-export * from "./memory-cache.js";
+
+// Re-export from common
+export { ICache, CacheOptions, MemoryCache } from "@opuspopuli/common";
+
+// Local implementations (Redis-specific, stays in extraction-provider)
 export * from "./redis-cache.js";
 export * from "./cache-factory.js";
